Support for some entities
Use isNight from BuildplateModel
Update example usage

+++ MCeToJava/Models/MCE/Entity.cs
@@ -6,4 +6,4 @@ using MathUtils.Vectors;
 
 namespace MCeToJava.Models.MCE;
 
-internal record Entity(string Name, float3 Position, float2 Rotation, float3 ShadowPosition, float ShadowSize, int OverlayColor, int ChangeColor, int MultiplicitiveTintChangeColor, Dictionary<string, object>? ExtraData, string SkinData, bool IsPersonaSkin);
+internal record Entity(string Name, double3 Position, float2 Rotation, float3 ShadowPosition, float ShadowSize, int OverlayColor, int ChangeColor, int MultiplicitiveTintChangeColor, Dictionary<string, object>? ExtraData, string SkinData, bool IsPersonaSkin);
